move twitter script logic into component
